requirements.txt
# Core dependencies from pyproject.toml
numpy>=1.24.0
librosa==0.11.0
s3tokenizer
torch==2.6.0
torchaudio==2.6.0
transformers==4.46.3
diffusers==0.29.0
resemble-perth==1.0.1
conformer==0.3.2
safetensors==0.5.3
spacy-pkuseg
pykakasi==2.3.0
pyloudnorm
omegaconf

# Web framework and utilities
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# Audio processing
soundfile>=0.12.1

# Production server
gunicorn==21.2.0
python-dotenv>=1.0.0

# Monitoring and logging
prometheus-client>=0.19.0

# Optional: For AWS deployment
boto3>=1.26.0
python-multipart>=0.0.6
